Modification Carte,Noeud,Segment

+++ src/Classe/Noeud.java
@@ -10,27 +10,27 @@ package Classe;
  */
 public class Noeud {
     
-    //Attributs
-    private int m_posX;
+   //Attributs
+    private float m_posX;
     
-    private int m_posY;
-
-    public int obtenir_posX() {
+    private float m_posY;
+    
+     public float obtenir_posX() {
         return m_posX;
     }
 
-    public int obtenir_posY() {
+    public float obtenir_posY() {
         return m_posY;
     }
     
     //Constructeur
-    Noeud(int p_posX, int p_posY)
+    Noeud(float p_posX, float p_posY)
     {
         m_posX = p_posX;
         m_posY = p_posY;
     }
     
-    boolean EstMemePosition(int p_posX, int p_posY)
+    boolean EstMemePosition(float p_posX, float p_posY)
     {
         return (m_posX == p_posX && m_posY == p_posY);
     }
